* CURD operations for doctors * Add the dtos, controllers and services for CRUD operations. * Add use cases * Design the Auth controller and doctor controller for operations. * Refactor naming * Handle register and login operations

+++ src/main/java/com/findmeadoc/application/dto/DoctorRegistrationResponse.java
@@ -1,12 +1,10 @@
 package com.findmeadoc.application.dto;
 
-import java.math.BigDecimal;
-
 public record DoctorRegistrationResponse(
         Long doctorId,
         String email,
         String name,
         String specialization,
-        BigDecimal consultationFee
+        Double consultationFee
 ) {
 }
